+++ modules/recast/SCsub
@@ -1,8 +1,9 @@
 #!/usr/bin/env python
 
 Import('env')
+Import('env_modules')
 
-# Not building in a separate env as core needs it
+env_recast = env_modules.Clone()
 
 # Thirdparty source files
 if env['builtin_recast']:
@@ -22,13 +23,10 @@ if env['builtin_recast']:
     ]
     thirdparty_sources = [thirdparty_dir + file for file in thirdparty_sources]
 
-    env.Append(CPPPATH=[thirdparty_dir, thirdparty_dir + "/Include"])
-
-    lib = env.add_library("recast_builtin", thirdparty_sources)
-    env.Append(LIBS=[lib])
+    env_recast.add_source_files(env.modules_sources, thirdparty_sources)
+    env_recast.Append(CPPPATH=[thirdparty_dir + "/Include"])
 
 # Godot source files
-env.add_source_files(env.modules_sources, "*.cpp")
-env.Append(CCFLAGS=['-DRECAST_ENABLED'])
+env_recast.add_source_files(env.modules_sources, "*.cpp")
 
 Export('env')
