+++ modules/recast/register_types.cpp
@@ -30,5 +30,10 @@
 
 #include "register_types.h"
 
-void register_recast_types() {}
+#include "navigation_mesh_editor_plugin.h"
+
+void register_recast_types() {
+	EditorPlugins::add_by_type<NavigationMeshEditorPlugin>();
+}
+
 void unregister_recast_types() {}
